+++ app/services/data_hub/register_school_importer/school_dry_run_creator.rb
@@ -1,0 +1,8 @@
+module DataHub
+  module RegisterSchoolImporter
+    class SchoolDryRunCreator < SchoolCreator
+      # No-op: simulate creation only
+      def create!(site, gias_school); end
+    end
+  end
+end
