--- app/services/register_school_importer/school_dry_run_creator.rb
@@ -1,6 +1,0 @@
-module RegisterSchoolImporter
-  class SchoolDryRunCreator < SchoolCreator
-    # No-op: simulate creation only
-    def create!(site, gias_school); end
-  end
-end
