+++ spec/services/data_hub/register_school_importer/import_summary_spec.rb
@@ -1,6 +1,6 @@
 require "rails_helper"
 
-RSpec.describe RegisterSchoolImporter::ImportSummary do
+RSpec.describe DataHub::RegisterSchoolImporter::ImportSummary do
   subject(:summary) { described_class.new }
 
   let(:provider_code) { "BU1" }
